+++ gendarme/rules/Gendarme.Rules.Naming/UseCorrectPrefixRule.cs
@@ -129,7 +129,7 @@ namespace Gendarme.Rules.Naming {
 				}
 			} else {
 				// class should _not_ look like 'CSomething" or like an interface 'IOops'
-				if (!IsCorrectTypeName (name)) { 
+				if (!name.Contains("@") && !IsCorrectTypeName (name)) { 
 					string s = String.Format (CultureInfo.InvariantCulture,
 						"The '{0}' type name starts with '{1}' prefix but, according to existing naming conventions, type names should not have any specific prefix.", 
 						name, name [0]);